+++ CHANGELOG.md
@@ -56,6 +56,13 @@ upgrades to `knex` package.
       full-table select queries in support of elasticsearch indexing.
     - Added `getMaxCumulusId` to `@cumulus/db` `BasePgModel` to allow all
       derived table classes to support querying the current max `cumulus_id`.
+  - **CUMULUS-2673**
+    - Added `elasticsearch_hostname` variable to
+      `postgres-migration-async-operation` lambda.
+    - Added `elasticsearch_security_group_id` to
+      `postgres-migration-async-operation` lambda.
+    - Added permission for `DyanmoDb:DeleteItem` to
+      `postgres-migration-async-operation` lambda.
 - Added `@cumulus/db/createRejectableTransaction()` to handle creating a Knex transaction that **will throw an error** if the transaction rolls back. [As of Knex 0.95+, promise rejection on transaction rollback is no longer the default behavior](https://github.com/knex/knex/blob/master/UPGRADING.md#upgrading-to-version-0950).
 - **CUMULUS-2670**
   - Updated `lambda_timeouts` string map variable for `cumulus` module to accept a